+++ modules/tun.py
@@ -15,6 +15,7 @@ import fcntl
 import os
 import select
 import struct
+import numpy as np
 
 # ioctl constants for Linux TUN/TAP
 _IFF_TUN   = 0x0001   # TUN device (no Ethernet header)
@@ -42,7 +43,7 @@ class TunDevice:
         self,
         name: str = "tun0",
         mtu: int = 1500,
-        poll_timeout: float = 0.05,
+        poll_timeout: np.float32 = np.float32(0.05),
     ) -> None:
         self.name = name
         self.mtu = mtu
